requirements.txt
fastapi
httpx
pytest
uvicorn
python-dotenv
psycopg2-binary
elasticsearch>=8,<9